yuzu: Fix truncation warnings within UI code

+++ src/yuzu/configuration/config.cpp
@@ -680,7 +680,7 @@ void Config::SaveValues() {
         qt_config->setValue("title_id", QVariant::fromValue<u64>(elem.first));
         qt_config->beginWriteArray("disabled");
         for (std::size_t j = 0; j < elem.second.size(); ++j) {
-            qt_config->setArrayIndex(j);
+            qt_config->setArrayIndex(static_cast<int>(j));
             qt_config->setValue("d", QString::fromStdString(elem.second[j]));
         }
         qt_config->endArray();
